change dtype of meta params

+++ src/dsbmm_bp/np_par_dsbmm_methods.py
@@ -575,7 +575,9 @@ class NumpyDSBMM:
             # lam_num[lam_num < TOL] = TOL
             # lam_den[lam_den < TOL] = 1.0
         # NB use relative rather than absolute difference here as lam could be large
-        tmp = np.divide(lam_num, lam_den, where=lam_den > 0, out=np.zeros_like(lam_num))
+        tmp = np.divide(
+            lam_num, lam_den, where=lam_den > 0, out=np.zeros_like(lam_num), dtype=float
+        )
         if not init:
             tmp = learning_rate * tmp + (1 - learning_rate) * self._lam
             tmp_diff = np.abs((tmp - self._lam) / self._lam).mean()
@@ -669,7 +671,11 @@ class NumpyDSBMM:
 
         # correct for numerical stability
         tmp = np.divide(
-            beta_num, beta_den, where=beta_den > 0, out=np.zeros_like(beta_num)
+            beta_num,
+            beta_den,
+            where=beta_den > 0,
+            out=np.zeros_like(beta_num),
+            dtype=float,
         )
         # tmp[tmp < TOL] = TOL
         # tmp[tmp > 1 - TOL] = 1 - TOL
@@ -749,7 +755,7 @@ class NumpyDSBMM:
             # xi[xi < TOL] = 1.0
             # zeta[zeta < TOL] = TOL
         # NB again use relative error here as could be large
-        tmp = np.divide(zeta, xi, where=xi > 0, out=np.zeros_like(zeta))
+        tmp = np.divide(zeta, xi, where=xi > 0, out=np.zeros_like(zeta), dtype=float)
         # tmp[tmp < TOL] = TOL
         if not init:
             tmp = learning_rate * tmp + (1 - learning_rate) * self.meta_params[s]
@@ -803,7 +809,7 @@ class NumpyDSBMM:
             )
             # xi[xi < TOL] = 1.0
             # rho[rho < TOL] = TOL
-        tmp = np.divide(rho, xi, where=xi > 0, out=np.zeros_like(rho))
+        tmp = np.divide(rho, xi, where=xi > 0, out=np.zeros_like(rho), dtype=float)
         # tmp[tmp < TOL] = TOL
         # tmp[tmp > 1 - TOL] = 1 - TOL
         if not init:
@@ -842,7 +848,7 @@ class NumpyDSBMM:
             )
             # rho[rho < TOL] = TOL
         xi = rho.sum(axis=-1, keepdims=True)
-        tmp = np.divide(rho, xi, where=xi > 0, out=np.zeros_like(rho))
+        tmp = np.divide(rho, xi, where=xi > 0, out=np.zeros_like(rho), dtype=float)
         # tmp[tmp < TOL] = TOL
         # tmp /= tmp.sum(axis=-1, keepdims=True)
         if not init:
